+++ src/windows/context_menu.rs
@@ -126,7 +126,7 @@ pub fn create_context_menu(app: &mut AppData) -> Task<Message>
     let backdrop_settings = NewLayerShellSettings
     {
         layer: Layer::Overlay,
-        size: Some((app.monitor_size.0 as u32, app.monitor_size.1 as u32)),
+        size: Some((app.monitor_size.0, app.monitor_size.1)),
         exclusive_zone: Some(0),
         keyboard_interactivity: KeyboardInteractivity::None,
         anchor: Anchor::Top | Anchor::Left,
